pyproject.toml
[project]
name = "tama-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "folium>=0.20.0",
    "geopandas>=1.1.1",
    "jupyter>=1.1.1",
    "mapclassify>=2.10.0",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "osmnx>=2.0.6",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
